fixed a few gcc-side warnings and a spelling error in the name of FormRecurrentLoop() that got fixed but partially undone during the last merge; added clear installation instructions for MS MPI SDK in the source file at the place where it fails if you don't have it

+++ Math/Math/QuantizedMatrix.cpp
@@ -170,7 +170,7 @@ namespace Microsoft { namespace MSR { namespace CNTK {
                     val = q.valQ.Unquantize(qVal);
                 }
 
-                fprintf(stderr, "%10d (%.10f)          \t", qVal, val);
+                fprintf(stderr, "%10d (%.10f)          \t", (int) qVal, val);
             }
             fprintf(stderr, "\n");
         }
